backend/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow-cpu
numpy<2.0.0
pillow
scikit-learn
matplotlib
requests
